% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{ApiAccess}
\alias{ApiAccess}
\title{ApiAccess}
\description{
/**
\itemize{
\item @param string $api_key       Your API key obtained from example.com
\item @param string $api_secret    Your API secret obtained from example.com
*/
ApiAccess class
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApiAccess$new()}}
\item \href{#method-config}{\code{ApiAccess$config()}}
\item \href{#method-get_api_url}{\code{ApiAccess$get_api_url()}}
\item \href{#method-set_path}{\code{ApiAccess$set_path()}}
\item \href{#method-get_config}{\code{ApiAccess$get_config()}}
\item \href{#method-endpoints}{\code{ApiAccess$endpoints()}}
\item \href{#method-send}{\code{ApiAccess$send()}}
\item \href{#method-nonce}{\code{ApiAccess$nonce()}}
\item \href{#method-auth}{\code{ApiAccess$auth()}}
\item \href{#method-post}{\code{ApiAccess$post()}}
\item \href{#method-get_endpoints}{\code{ApiAccess$get_endpoints()}}
\item \href{#method-decode}{\code{ApiAccess$decode()}}
\item \href{#method-clone}{\code{ApiAccess$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$new(api_key, api_secret, api_version = "v1")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-config"></a>}}
\if{latex}{\out{\hypertarget{method-config}{}}}
\subsection{Method \code{config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$config(arr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_api_url"></a>}}
\if{latex}{\out{\hypertarget{method-get_api_url}{}}}
\subsection{Method \code{get_api_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$get_api_url(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_path"></a>}}
\if{latex}{\out{\hypertarget{method-set_path}{}}}
\subsection{Method \code{set_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$set_path(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_config}{}}}
\subsection{Method \code{get_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-endpoints"></a>}}
\if{latex}{\out{\hypertarget{method-endpoints}{}}}
\subsection{Method \code{endpoints()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$endpoints(input = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send"></a>}}
\if{latex}{\out{\hypertarget{method-send}{}}}
\subsection{Method \code{send()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$send(path, input, payload)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nonce"></a>}}
\if{latex}{\out{\hypertarget{method-nonce}{}}}
\subsection{Method \code{nonce()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$nonce()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auth"></a>}}
\if{latex}{\out{\hypertarget{method-auth}{}}}
\subsection{Method \code{auth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$auth()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post"></a>}}
\if{latex}{\out{\hypertarget{method-post}{}}}
\subsection{Method \code{post()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$post(input, payload)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_endpoints"></a>}}
\if{latex}{\out{\hypertarget{method-get_endpoints}{}}}
\subsection{Method \code{get_endpoints()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$get_endpoints()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decode"></a>}}
\if{latex}{\out{\hypertarget{method-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$decode(result)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAccess$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
