% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.r
\docType{class}
\name{buyer-class}
\alias{buyer-class}
\alias{buyer}
\title{Buyer class for interacting with the network}
\arguments{
\item{api_key}{The \verb{API KEY} from the website console}

\item{ip_addr}{The \verb{IP Address} key from the website console}

\item{port}{\link{optional} The \code{PORT} from the website console}

\item{use_https}{the https boolean toogle}
}
\description{
Buyer class for interacting with the network
}
\section{Methods}{

\describe{
\item{\code{get_columns(query_key = NULL)}}{Return the table ddl columns}

\item{\code{get_key()}}{Returns a subkey to interact with the network}

\item{\code{get_key_list()}}{Return the list of sub keys for the current buyer}

\item{\code{print_query_history()}}{Pretty print the history queries with accuracy and results}

\item{\code{query(query_key = NULL, query = "")}}{returns result and accuracy of the query}
}}

