% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.r
\name{get_columns}
\alias{get_columns}
\title{Return the table ddl columns}
\usage{
get_columns(buyer, query_key)
}
\arguments{
\item{buyer}{The \code{Buyer} object from the \code{new_buyer} function}

\item{query_key}{\link{optional} subkey query key}
}
\value{
returns the list of columns
#'
}
\description{
Return the table ddl columns
}
