% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.r
\name{get_key}
\alias{get_key}
\title{Returns a sub key for the passed in buyer}
\usage{
get_key(buyer)
}
\arguments{
\item{buyer}{The \code{Buyer} object from the \code{new_buyer} function}
}
\value{
the subkey to interact with the network
}
\description{
Returns a sub key for the passed in buyer
}
