% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.r
\name{get_key_list}
\alias{get_key_list}
\title{Return the list of sub keys for the passed in buyer}
\usage{
get_key_list(buyer)
}
\arguments{
\item{buyer}{The \code{Buyer} object from the \code{new_buyer} function}
}
\value{
the list of sub keys
#'
}
\description{
Return the list of sub keys for the passed in buyer
}
