% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.r
\name{new_buyer}
\alias{new_buyer}
\title{Instantiate a new \code{Buyer} object}
\usage{
new_buyer(api_key, ip_addr, port = "8080")
}
\arguments{
\item{api_key}{The \verb{API KEY} from the website console}

\item{ip_addr}{The \verb{IP Address} key from the website console}

\item{port}{\link{optional} The \code{PORT} from the website console}
}
\value{
\code{Buyer} object to interact with the network
}
\description{
Instantiate a new \code{Buyer} object
}
\examples{
api_key <- "buyer_key"
ip_addr <- "127.0.0.1"
buyer <- new_buyer(api_key, ip_addr)

}
