% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.r
\name{print_query_history}
\alias{print_query_history}
\title{Pretty print the history queries with accuracy and results}
\usage{
print_query_history(buyer)
}
\arguments{
\item{buyer}{The \code{Buyer} object from the \code{new_buyer} function}
}
\description{
Pretty print the history queries with accuracy and results
}
