% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.r
\name{query}
\alias{query}
\title{Return the list of sub keys for the passed in buyer}
\usage{
query(buyer, query_key = NULL, query)
}
\arguments{
\item{buyer}{The \code{Buyer} object from the \code{new_buyer} function}

\item{query_key}{\link{optional} subkey query key}

\item{query}{query string}
}
\value{
returns \code{result} and \code{accuracy} of the query
#' @examples
sql_query <- "select count(*) as numpeople from public.condition_era_death"
result <- query(buyer, query = sql_query)
print_query_history(buyer)
#'
}
\description{
Return the list of sub keys for the passed in buyer
}
