% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdk-package.R
\docType{package}
\name{rsdk-package}
\alias{rsdk-package}
\alias{_PACKAGE}
\title{rsdk: Fortress SDK}
\description{
SDK for accesssing and performing private computation on Fortress platform.
}
\author{
\strong{Maintainer}: Fortress Labs \email{something@fortresslabs.com}

}
\keyword{internal}
