% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_replace_symbols}
\alias{str_replace_symbols}
\title{Replace Symbols}
\usage{
str_replace_symbols(str)
}
\arguments{
\item{str}{a character vector.}
}
\value{

}
\description{
Replaces '+' to '-', '/' to '_' and '=' to ''
}
\examples{
str_replace_symbols("a+b-c/d_e=f")
}
